+++ spec/services/create_programme_activity_spec.rb
@@ -11,6 +11,10 @@ RSpec.describe CreateProgrammeActivity do
       expect(result.organisation).to eq(organisation)
     end
 
+    it "saves the reporting organisation reference" do
+      expect(result.reporting_organisation_reference).to eq(organisation.iati_reference)
+    end
+
     it "sets the parent Activity to the fund" do
       expect(result.activity).to eq(fund)
     end