+++ CHANGELOG.md
@@ -55,3 +55,4 @@
 - Users can view implementing organisations in the Activity XML
 - Planned start and end dates are mandatory
 - Actual start and end dates must not be in the future
+- Activity XML includes 'iati-identifier' which is includes the reporting organisation